[002_material_class_tests], issue #2, moved data setup to conftest file
