+++ tests/test_data/material_class_data.json
@@ -45,10 +45,10 @@
     "expected_from_corr": {
         "density": [
             229.2,
-            309.06666666666666, 
-            469.03999999999996, 
+            309.0667, 
+            469.04, 
             709.0285714285715, 
-            1029.0222222222221
+            1029.0222
         ],
         "thermal_conductivity": [
             113.779, 
@@ -59,8 +59,8 @@
         ],
         "heat_capacity": [
             1.2784, 
-            1.4104444444444444, 
-            1.3337279999999998, 
+            1.4104444, 
+            1.333728, 
             1.2775918367346937, 
             1.254004938271605    
         ],
@@ -317,19 +317,19 @@
             0.0024358,
             0.002089184,
             0.0018411185,
-            0.0016560605000000002,
+            0.0016560605,
             0.001513399,
             0.0014004605,
-            0.0013090735000000002,
+            0.0013090735,
             0.0012337585,
             0.001170717,
             0.001117241,
             0.0010713525,
-            0.0010315749999999999,
-            0.0009967865000000001,
-            0.0009661208000000001,
+            0.001031575,
+            0.0009967865,
+            0.0009661208,
             0.0009388979,
-            0.0009145780999999999
+            0.0009145781
         ],
         "thermal_conductivity": [
             9.73847,
@@ -783,7 +783,7 @@
             136.70973868935096,
             136.97868645110097,
             137.3225282229311,
-            137.74150666666665,
+            137.7415067,
             138.2358333445775
         ],
         "viscosity": [
@@ -805,17 +805,17 @@
         ],
         "thermal_conductivity": [
             16.625,
-            17.174999999999997,
+            17.175,
             17.725,
             18.275,
             18.825,
             19.375,
-            19.924999999999997,
-            20.474999999999998,
+            19.925,
+            20.475,
             21.025,
             21.575,
             22.125,
-            22.674999999999997
+            22.675
         ]
     },
     "bismuth_corr" : {
@@ -888,14 +888,14 @@
             0.0008423169617084873
         ],
         "thermal_conductivity": [
-            12.802499999999998,
+            12.8025,
             13.2775,
             13.7525,
             14.2275,
             14.7025,
-            15.177499999999998,
+            15.1775,
             15.6525,
-            16.127499999999998
+            16.1275
         ]
     },
     "lbe_corr" : {
@@ -986,14 +986,14 @@
             10.444684375,
             11.137934375,
             11.819659375,
-            12.489859374999998,
+            12.489859375,
             13.148534375,
             13.795684375,
-            14.431309375000001,
+            14.431309375,
             15.055409375,
             15.667984375,
             16.269034375,
-            16.858559375000002,
+            16.858559375,
             17.436559375,
             18.003034375,
             18.557984375
@@ -1065,10 +1065,10 @@
             25.42875,
             25.80375,
             26.06875,
-            26.223750000000003,
+            26.22375,
             26.26875,
             26.20375,
-            26.028750000000002,
+            26.02875,
             25.74375,
             25.34875,
             24.84375,
